+++ src/components/graph/selectionToolbox/HelpButton.vue
@@ -18,17 +18,17 @@ import Button from 'primevue/button'
 import { computed } from 'vue'
 
 import { useNodeLibrarySidebarTab } from '@/composables/sidebarTabs/useNodeLibrarySidebarTab'
-import { useNodeHelp } from '@/composables/useNodeHelp'
 import { useCanvasStore } from '@/stores/graphStore'
 import type { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
+import { useNodeHelpStore } from '@/stores/workspace/nodeHelpStore'
 import { useSidebarTabStore } from '@/stores/workspace/sidebarTabStore'
 import { isLGraphNode } from '@/utils/litegraphUtil'
 
 const canvasStore = useCanvasStore()
 const nodeDefStore = useNodeDefStore()
 const sidebarTabStore = useSidebarTabStore()
-const { openHelp } = useNodeHelp()
+const nodeHelpStore = useNodeHelpStore()
 const { id: nodeLibraryTabId } = useNodeLibrarySidebarTab()
 
 const nodeDef = computed<ComfyNodeDefImpl | null>(() => {
@@ -44,6 +44,6 @@ const showHelp = () => {
   if (sidebarTabStore.activeSidebarTabId !== nodeLibraryTabId) {
     sidebarTabStore.toggleSidebarTab(nodeLibraryTabId)
   }
-  openHelp(def)
+  nodeHelpStore.openHelp(def)
 }
 </script>
