+++ tests-ui/tests/store/nodeHelpStore.test.ts
@@ -1,8 +1,9 @@
 import { flushPromises } from '@vue/test-utils'
+import { createPinia, setActivePinia } from 'pinia'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { nextTick } from 'vue'
 
-import { useNodeHelp } from '@/composables/useNodeHelp'
+import { useNodeHelpStore } from '@/stores/workspace/nodeHelpStore'
 
 vi.mock('@/scripts/api', () => ({
   api: {
@@ -66,7 +67,7 @@ vi.mock('marked', () => ({
   }
 }))
 
-describe('useNodeHelp', () => {
+describe('nodeHelpStore', () => {
   // Define a mock node for testing
   const mockCoreNode = {
     name: 'TestNode',
@@ -91,133 +92,131 @@ describe('useNodeHelp', () => {
   global.fetch = mockFetch
 
   beforeEach(() => {
+    // Setup Pinia
+    setActivePinia(createPinia())
     mockFetch.mockReset()
   })
 
-  // Reset state before each test
-  beforeEach(() => {
-    const { closeHelp } = useNodeHelp()
-    closeHelp()
-  })
-
   it('should initialize with empty state', () => {
-    const { currentHelpNode, isHelpOpen } = useNodeHelp()
-    expect(currentHelpNode.value).toBeNull()
-    expect(isHelpOpen.value).toBe(false)
+    const nodeHelpStore = useNodeHelpStore()
+    expect(nodeHelpStore.currentHelpNode).toBeNull()
+    expect(nodeHelpStore.isHelpOpen).toBe(false)
   })
 
   it('should open help for a node', () => {
-    const { currentHelpNode, isHelpOpen, openHelp } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
 
-    expect(currentHelpNode.value).toStrictEqual(mockCoreNode)
-    expect(isHelpOpen.value).toBe(true)
+    expect(nodeHelpStore.currentHelpNode).toStrictEqual(mockCoreNode)
+    expect(nodeHelpStore.isHelpOpen).toBe(true)
   })
 
   it('should close help', () => {
-    const { currentHelpNode, isHelpOpen, openHelp, closeHelp } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
-    openHelp(mockCoreNode as any)
-    expect(isHelpOpen.value).toBe(true)
+    nodeHelpStore.openHelp(mockCoreNode as any)
+    expect(nodeHelpStore.isHelpOpen).toBe(true)
 
-    closeHelp()
-    expect(currentHelpNode.value).toBeNull()
-    expect(isHelpOpen.value).toBe(false)
+    nodeHelpStore.closeHelp()
+    expect(nodeHelpStore.currentHelpNode).toBeNull()
+    expect(nodeHelpStore.isHelpOpen).toBe(false)
   })
 
   it('should generate correct baseUrl for core nodes', async () => {
-    const { openHelp, baseUrl } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await nextTick()
 
-    expect(baseUrl.value).toBe(`/docs/${mockCoreNode.name}/`)
+    expect(nodeHelpStore.baseUrl).toBe(`/docs/${mockCoreNode.name}/`)
   })
 
   it('should generate correct baseUrl for custom nodes', async () => {
-    const { openHelp, baseUrl } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await nextTick()
 
-    expect(baseUrl.value).toBe('/extensions/test_module/docs/')
+    expect(nodeHelpStore.baseUrl).toBe('/extensions/test_module/docs/')
   })
 
   it('should render markdown content correctly', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '# Test Help\nThis is test help content'
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
 
-    expect(renderedHelpHtml.value).toContain('This is test help content')
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
+      'This is test help content'
+    )
   })
 
   it('should handle fetch errors and fall back to description', async () => {
-    const { openHelp, renderedHelpHtml, error } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: false,
       statusText: 'Not Found'
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
 
-    expect(error.value).toBe('Not Found')
-    expect(renderedHelpHtml.value).toContain(mockCoreNode.description)
+    expect(nodeHelpStore.error).toBe('Not Found')
+    expect(nodeHelpStore.renderedHelpHtml).toContain(mockCoreNode.description)
   })
 
   it('should include alt attribute for images', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '![image](test.jpg)'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain('alt="image"')
+    expect(nodeHelpStore.renderedHelpHtml).toContain('alt="image"')
   })
 
   it('should prefix relative video src in custom nodes', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '<video src="video.mp4"></video>'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       'src="/extensions/test_module/docs/video.mp4"'
     )
   })
 
   it('should prefix relative video src for core nodes with node-specific base URL', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '<video src="video.mp4"></video>'
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/video.mp4"`
     )
   })
 
   it('should prefix relative source src in custom nodes', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
@@ -225,15 +224,15 @@ describe('useNodeHelp', () => {
         '<video><source src="video.mp4" type="video/mp4" /></video>'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       'src="/extensions/test_module/docs/video.mp4"'
     )
   })
 
   it('should prefix relative source src for core nodes with node-specific base URL', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
@@ -241,26 +240,26 @@ describe('useNodeHelp', () => {
         '<video><source src="video.webm" type="video/webm" /></video>'
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/video.webm"`
     )
   })
 
   it('should handle loading state', async () => {
-    const { openHelp, isLoading } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockImplementationOnce(() => new Promise(() => {})) // Never resolves
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await nextTick()
 
-    expect(isLoading.value).toBe(true)
+    expect(nodeHelpStore.isLoading).toBe(true)
   })
 
   it('should try fallback URL for custom nodes', async () => {
-    const { openHelp } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch
       .mockResolvedValueOnce({
@@ -272,7 +271,7 @@ describe('useNodeHelp', () => {
         text: async () => '# Fallback content'
       })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
 
     expect(mockFetch).toHaveBeenCalledTimes(2)
@@ -285,53 +284,55 @@ describe('useNodeHelp', () => {
   })
 
   it('should prefix relative img src in raw HTML for custom nodes', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '# Test\n<img src="image.png" alt="Test image">'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       'src="/extensions/test_module/docs/image.png"'
     )
-    expect(renderedHelpHtml.value).toContain('alt="Test image"')
+    expect(nodeHelpStore.renderedHelpHtml).toContain('alt="Test image"')
   })
 
   it('should prefix relative img src in raw HTML for core nodes', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '# Test\n<img src="image.png" alt="Test image">'
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/image.png"`
     )
-    expect(renderedHelpHtml.value).toContain('alt="Test image"')
+    expect(nodeHelpStore.renderedHelpHtml).toContain('alt="Test image"')
   })
 
   it('should not prefix absolute img src in raw HTML', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
       text: async () => '<img src="/absolute/image.png" alt="Absolute">'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain('src="/absolute/image.png"')
-    expect(renderedHelpHtml.value).toContain('alt="Absolute"')
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
+      'src="/absolute/image.png"'
+    )
+    expect(nodeHelpStore.renderedHelpHtml).toContain('alt="Absolute"')
   })
 
   it('should not prefix external img src in raw HTML', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
@@ -339,16 +340,16 @@ describe('useNodeHelp', () => {
         '<img src="https://example.com/image.png" alt="External">'
     })
 
-    openHelp(mockCustomNode as any)
+    nodeHelpStore.openHelp(mockCustomNode as any)
     await flushPromises()
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       'src="https://example.com/image.png"'
     )
-    expect(renderedHelpHtml.value).toContain('alt="External"')
+    expect(nodeHelpStore.renderedHelpHtml).toContain('alt="External"')
   })
 
   it('should handle various quote styles in media src attributes', async () => {
-    const { openHelp, renderedHelpHtml } = useNodeHelp()
+    const nodeHelpStore = useNodeHelpStore()
 
     mockFetch.mockResolvedValueOnce({
       ok: true,
@@ -369,29 +370,29 @@ Testing quote styles in properly formed HTML:
 The MEDIA_SRC_REGEX handles both single and double quotes in img, video and source tags.`
     })
 
-    openHelp(mockCoreNode as any)
+    nodeHelpStore.openHelp(mockCoreNode as any)
     await flushPromises()
 
     // Check that all media elements with different quote styles are prefixed correctly
     // Double quotes remain as double quotes
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/video1.mp4"`
     )
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/image1.png"`
     )
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src="/docs/${mockCoreNode.name}/video3.mp4"`
     )
 
     // Single quotes remain as single quotes in the output
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src='/docs/${mockCoreNode.name}/video2.mp4'`
     )
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src='/docs/${mockCoreNode.name}/image2.png'`
     )
-    expect(renderedHelpHtml.value).toContain(
+    expect(nodeHelpStore.renderedHelpHtml).toContain(
       `src='/docs/${mockCoreNode.name}/video3.webm'`
     )
   })
